Update references to recently moved internal methods
