+++ src/lib/index.js
@@ -1,15 +1,9 @@
-let checkRuntimes = require('./check-runtimes')
-let env = require('./env')
 let getFlags = require('./flags')
-let maybeHydrate = require('./maybe-hydrate')
 let runtimeEval = require('./runtime-eval')
 let template = require('./template')
 
 module.exports = {
-  checkRuntimes,
-  env,
   getFlags,
-  maybeHydrate,
   runtimeEval,
   template
 }
